requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
async-timeout==4.0.2
black==22.10.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==4.4.2
Deprecated==1.2.13
docopt==0.6.2
Flask==2.3.2
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
imageio==2.22.3
imageio-ffmpeg==0.4.7
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
moviepy==1.0.3
mutagen==1.46.0
mypy-extensions==0.4.3
npm==0.1.1
numpy==1.23.4
optional-django==0.1.0
packaging==21.3
pathspec==0.10.1
Pillow==9.3.0
pipreqs==0.4.13
platformdirs==2.5.2
proglog==0.1.10
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.13.0
pyparsing==3.0.9
python-dotenv==1.0.0
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
redis==4.3.4
requests==2.28.1
rich==12.6.0
rsa==4.9
six==1.16.0
spotipy==2.21.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.3.6
wrapt==1.14.1
yarg==0.1.9
zipp==3.16.2
